Move hasGetter/SetterBody to utility class
